+++ cli/notifymanager.py
@@ -5,6 +5,8 @@ __version__ = "3.2.4"
 
 import sys
 import os
+import argparse
+import textwrap
 from datetime import datetime
 
 dir_lib = os.path.dirname(os.path.realpath(__file__)) + '/../lib'
@@ -23,7 +25,7 @@ class NotifierCLI(GenericCLI, ControllerCLI):
     class ComplNProto(object):
 
         def __call__(self, prefix, **kwargs):
-            return ['mqtt:', 'json:', 'db:', 'influxdb:'] if \
+            return ['mqtt:', 'json:', 'db:', 'influxdb:', 'prometheus:'] if \
                     prefix.find(':') == -1 else True
 
     class ComplN(ComplGeneric):
@@ -62,13 +64,16 @@ class NotifierCLI(GenericCLI, ControllerCLI):
 
     def add_notifier_common_functions(self):
         ap_list = self.sp.add_parser('list', help='List notifiers')
-        ap_create = self.sp.add_parser('create', help='Create notifier')
+        ap_create = self.sp.add_parser('create', help='Create notifier',
+            formatter_class=argparse.RawTextHelpFormatter)
         ap_create.add_argument('i', help='Notifier ID', metavar='ID')
-        ap_create.add_argument('p', help='Notifier properties: ' + \
-                'json:http(s)://[key]@uri[#jsonrpc] or ' + \
-                'mqtt:[username:password]@host:[port] or ' + \
-                'db:db_uri or ' + \
-                'influxdb:http(s)://uri#database',
+        ap_create.add_argument('p', help=textwrap.dedent('''
+        Notifier properties:
+            json:http(s)://[key]@uri[#jsonrpc]
+            mqtt:[username:password]@host:[port]
+            db:db_uri
+            influxdb:http(s)://uri#database
+            prometheus:'''),
                 metavar='PROPS').completer = self.ComplNProto()
         ap_create.add_argument(
             '-s',
@@ -302,6 +307,8 @@ class NotifierCLI(GenericCLI, ControllerCLI):
             db_uri = ':'.join(p[1:])
             n = eva.notify.SQLANotifier(
                 notifier_id=notifier_id, db_uri=db_uri, keep=None, space=space)
+        elif p[0] == 'prometheus':
+            n = eva.notify.PrometheusNotifier(notifier_id = notifier_id)
         else:
             self.print_err('notifier type unknown %s' % p[0])
             return self.local_func_result_failed